+++ examples/temperature.rs
@@ -13,7 +13,6 @@
 //! Rust Phidget example application to read temperature.
 //!
 
-use clap::{arg, value_parser, ArgAction};
 use phidget::Phidget;
 use std::{thread, time::Duration};
 
@@ -26,45 +25,16 @@ const VERSION: &str = env!("CARGO_PKG_VERSION");
 // --------------------------------------------------------------------------
 
 fn main() -> anyhow::Result<()> {
-    let opts = clap::Command::new("temperature")
-        .version(VERSION)
-        .author(env!("CARGO_PKG_AUTHORS"))
-        .about("Phidget Temperature Monitoring Example")
-        .disable_help_flag(true)
-        .arg(
-            arg!(--help "Print help information")
-                .short('?')
-                .action(ArgAction::Help),
-        )
-        .arg(
-            arg!(-s --serial [serial_num] "Specify the serial number of the device to open")
-                .value_parser(value_parser!(i32)),
-        )
-        .arg(
-            arg!(-c --channel [chan] "Specify the channel number of the device to open")
-                .value_parser(value_parser!(i32)),
-        )
-        .arg(
-            arg!(-p --port [port] "Use a specific port on a VINT hub directly")
-                .value_parser(value_parser!(i32)),
-        )
-        .get_matches();
-
+    println!("Phidgets-rs {VERSION}");
+    let port = 0; // Use a specific port on a VINT hub directly
+    let serial = 0; // Specify the serial number of the device to open
+    let channel = 0; // Specify the channel number of the device to open
     println!("Opening Phidget temperature sensor...");
-    let mut sensor = phidget::TemperatureSensor::new();
+    let mut sensor = phidget::devices::TemperatureSensor::new();
 
-    // Some device selection filters...
-    if let Some(&port) = opts.get_one::<i32>("port") {
-        sensor.set_hub_port(port)?;
-    }
-
-    if let Some(&num) = opts.get_one::<i32>("serial") {
-        sensor.set_serial_number(num)?;
-    }
-
-    if let Some(&chan) = opts.get_one::<i32>("channel") {
-        sensor.set_channel(chan)?;
-    }
+    sensor.set_hub_port(port)?;
+    sensor.set_serial_number(serial)?;
+    sensor.set_channel(channel)?;
 
     sensor.open_wait(TIMEOUT)?;