+++ examples/voltage_in.rs
@@ -12,7 +12,6 @@
 
 //! Rust Phidget example application to read voltage input values.
 
-use clap::{arg, value_parser, ArgAction};
 use phidget::Phidget;
 use std::{thread, time::Duration};
 
@@ -25,51 +24,24 @@ const VERSION: &str = env!("CARGO_PKG_VERSION");
 // --------------------------------------------------------------------------
 
 fn main() -> anyhow::Result<()> {
-    let opts = clap::Command::new("voltage_in")
-        .version(VERSION)
-        .author(env!("CARGO_PKG_AUTHORS"))
-        .about("Phidget Analog Input Example")
-        .disable_help_flag(true)
-        .arg(
-            arg!(--help "Print help information")
-                .short('?')
-                .action(ArgAction::Help),
-        )
-        .arg(
-            arg!(-s --serial [serial_num] "Specify the serial number of the device to open")
-                .value_parser(value_parser!(i32)),
-        )
-        .arg(
-            arg!(-c --channel [chan] "Specify the channel number of the device to open")
-                .value_parser(value_parser!(i32)),
-        )
-        .arg(
-            arg!(-p --port [port] "Use a specific port on a VINT hub directly")
-                .value_parser(value_parser!(i32)),
-        )
-        .arg(arg!(-h --hub "Use a hub VINT input port directly").action(ArgAction::SetTrue))
-        .get_matches();
+    println!("Phidgets-rs {VERSION}");
 
-    let use_hub = opts.get_flag("hub");
+    let port = 0; // Use a specific port on a VINT hub directly
+    let serial = 0; // Specify the serial number of the device to open
+    let channel = 0; // Specify the channel number of the device to open
+
+
+    let use_hub = true;
 
     println!("Opening Phidget voltage input device...");
-    let mut vin = phidget::VoltageInput::new();
+    let mut vin = phidget::devices::VoltageInput::new();
 
     // Whether we should use a hub port directly as the input,
     // and if so, which one?
     vin.set_is_hub_port_device(use_hub)?;
-    if let Some(&port) = opts.get_one::<i32>("port") {
-        vin.set_hub_port(port)?;
-    }
-
-    // Some other device selection filters...
-    if let Some(&num) = opts.get_one::<i32>("serial") {
-        vin.set_serial_number(num)?;
-    }
-
-    if let Some(&chan) = opts.get_one::<i32>("channel") {
-        vin.set_channel(chan)?;
-    }
+    vin.set_hub_port(port)?;
+    vin.set_serial_number(serial)?;
+    vin.set_channel(channel)?;
 
     vin.open_wait(TIMEOUT)?;