+++ src/bin/phidget.rs
@@ -7,7 +7,7 @@ fn main() -> anyhow::Result<()> {
     println!("{}", phidget::library_version()?);
     println!("{}", phidget::library_version_number()?);
 
-    let mut hum_sensor = phidget::HumiditySensor::new();
+    let mut hum_sensor = phidget::devices::HumiditySensor::new();
     phidget::phidget::set_on_attach_handler(&mut hum_sensor, |_| {
         println!("Humidity sensor attached");
     })?;
@@ -30,7 +30,7 @@ fn main() -> anyhow::Result<()> {
     let humidity = hum_sensor.humidity()?;
     println!("Humidity: {}", humidity);
 
-    let mut temp_sensor = phidget::TemperatureSensor::new();
+    let mut temp_sensor = phidget::devices::TemperatureSensor::new();
     phidget::phidget::set_on_attach_handler(&mut temp_sensor, |_| {
         println!("Temperature sensor attached");
     })?;
@@ -53,12 +53,12 @@ fn main() -> anyhow::Result<()> {
     let temperature = temp_sensor.temperature()?;
     println!("Temperature: {}\n", temperature);
 
-    hum_sensor.set_on_humidity_change_handler(|_s: &phidget::HumiditySensor, humidity: f64| {
+    hum_sensor.set_on_humidity_change_handler(|_s: &phidget::devices::HumiditySensor, humidity: f64| {
         println!("Humidity: {}", humidity);
     })?;
 
     temp_sensor.set_on_temperature_change_handler(
-        |_s: &phidget::TemperatureSensor, temperature: f64| {
+        |_s: &phidget::devices::TemperatureSensor, temperature: f64| {
             println!("Temerature: {}", temperature);
         },
     )?;