+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log for phidget-rs library crate
 
+
+## [v0.1.2]  - unreleased
+
+- Added stepper implementation. Still missing async wrapper
+- Seperated digital-io to digital-input and ouput. Completion of available functions from binding.
+- Removed clap as it is not necessary to have it to use the code and it adds bloat for the examples which should be the minimum code required to use the crate.
+- Moved devices to specific modul to keep the root directory clean
+
+## [v0.1.1](https://github.com/fpagliughi/phidget-rs/compare/v0.1.0..v0.1.1)  - 2023-04-20
+
+- Added attach/detach setup functions for the devices
+- Removed generic callback setters from `Phidget` trait to allow it to be used as an object, like `Vec<Box<Phidget>>`
+- Phidgets implement `Send` trait
+- Added a `GenericPhidget` type
+- Downgraded clap to v3.2 (from 4.x) to allow for lower/older MRSV (now Rust v0.59.0)
+
 ## [v0.1.0](https://github.com/fpagliughi/phidget-rs/tree/v0.1.0) - 2023-04-19
 
 Initial release with basic support for the following sensors: